arch/x86_64: fix halt in lapic_timer_calc_freq_bsp when building with -O3
The problem was that IOAPIC accesses were not volatile and the function
called ioapic_get_vector_by_irq, which gets the amount of pins of each
IOAPIC by reading it's register.

The pit_irq function has also been marked noinline so it can be properly
used as an interrupt handler.